Empyrean Web/Empyrean-Web-master/requirements.txt
alembic==1.1.0
blinker==1.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.1
psycopg2==2.8.3
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
WTForms==2.2.1
